TST: add tests for unversioned interface
