+++ tests/conftest.py
@@ -12,11 +12,15 @@ import audbackend
 
 # list of backends that will be tested by default
 pytest.BACKENDS = [
-    # 'artifactory',
+    'artifactory',
     'file-system',
 ]
 if os.name != 'nt':
     pytest.BACKENDS.append('single-folder')
+
+
+pytest.VERSIONED = [(backend, True) for backend in pytest.BACKENDS]
+pytest.UNVERSIONED = [(backend, False) for backend in pytest.BACKENDS]
 
 # UID for test session
 # Repositories on the host will be named
@@ -59,11 +63,11 @@ def owner(request):
 @pytest.fixture(scope='function', autouse=False)
 def backend(hosts, request):
 
-    name = request.param
+    name, versioned = request.param
     host = hosts[name]
     repository = f'unittest-{pytest.UID}-{audeer.uid()[:8]}'
 
-    backend = audbackend.create(name, host, repository)
+    backend = audbackend.create(name, host, repository, versioned=versioned)
 
     yield backend
